OAI-61: Moved currently used converters to legacy submodule
