+++ claim_ai/evaluation/converter.py
@@ -4,7 +4,7 @@ from typing import List
 import traceback
 import logging
 
-from claim_ai.evaluation.converters import AiConverter
+from claim_ai.evaluation.converters.legacy_converter import AiConverter
 from claim_ai.evaluation.evaluation_result import EvaluationResult
 from claim_ai.apps import ClaimAiConfig
