+++ claim_ai/tests/test_ai_output_converter.py
@@ -6,7 +6,7 @@ from unittest import mock
 
 from claim_ai.apps import ClaimAiConfig
 from claim_ai.evaluation import converter
-from claim_ai.evaluation.converters import AiConverter
+from claim_ai.evaluation.converters.legacy_converter import AiConverter
 from claim_ai.evaluation.evaluation_result import EvaluationResult
 from claim_ai.tests import socket_data, AiConvertionHelper
